+++ app/src/main/kotlin/org/virtuslab/bazelsteward/app/MavenDependencyKind.kt
@@ -1,5 +1,6 @@
 package org.virtuslab.bazelsteward.app
 
+import mu.KotlinLogging
 import org.virtuslab.bazelsteward.core.library.Library
 import org.virtuslab.bazelsteward.core.library.Version
 import org.virtuslab.bazelsteward.core.replacement.Heuristic
@@ -10,6 +11,8 @@ import org.virtuslab.bazelsteward.maven.MavenLibraryId
 import org.virtuslab.bazelsteward.maven.MavenRepository
 import java.nio.file.Path
 
+private val logger = KotlinLogging.logger {}
+
 class MavenDependencyKind(
   private val mavenDataExtractor: MavenDataExtractor,
   private val mavenRepository: MavenRepository
@@ -19,6 +22,8 @@ class MavenDependencyKind(
 
   override suspend fun findAvailableVersions(workspaceRoot: Path): Map<Library, List<Version>> {
     val data = mavenDataExtractor.extract()
+    logger.debug { "Repositories " + data.repositories.toString() }
+    logger.debug { "Dependencies: " + data.dependencies.map { it.id.name + " " + it.version.value }.toString() }
     return mavenRepository.findVersions(data) as Map<Library, List<Version>>
   }
