--- app/src/test/kotlin/org/virtuslab/bazelsteward/app/MainTest.kt
@@ -1,10 +1,0 @@
-package org.virtuslab.bazelsteward.app
-
-import org.junit.jupiter.api.Test
-
-class MainTest {
-  @Test
-  fun `should run`() {
-    // App.main(arrayOf())
-  }
-}
