+++ bazel/src/main/kotlin/org/virtuslab/bazelsteward/rules/BazelRulesExtractor.kt
@@ -9,7 +9,6 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import mu.KotlinLogging
 import org.virtuslab.bazelsteward.core.AppConfig
-import org.virtuslab.bazelsteward.core.common.BazelFileSearch
 import org.virtuslab.bazelsteward.core.common.CommandRunner
 import org.virtuslab.bazelsteward.core.library.SimpleVersion
 import org.virtuslab.bazelsteward.core.rules.RuleLibrary
@@ -38,7 +37,7 @@ class BazelRulesExtractor(private val appConfig: AppConfig) {
     withContext(Dispatchers.IO) {
       val dumpRepositoriesContent = javaClass.classLoader.getResource("bazel/resources/dump_repositories.bzlignore")?.readText()
         ?: throw RuntimeException("Could not find dump_repositories template, which is required for detecting used bazel repositories")
-      val tempFileForBzl = createTempFile(directory = appConfig.path, suffix = ".bzl").toFile()
+      val tempFileForBzl = createTempFile(directory = appConfig.workspaceRoot, suffix = ".bzl").toFile()
       tempFileForBzl.appendText(dumpRepositoriesContent)
 
       val workspaceFilePath = listOf("WORKSPACE.bazel", "WORKSPACE")
@@ -55,10 +54,10 @@ class BazelRulesExtractor(private val appConfig: AppConfig) {
         |)""".trimMargin()
       )
       // solution from https://github.com/bazelbuild/bazel/issues/6377#issuecomment-1237791008
-      CommandRunner.run("bazel build @all_external_repositories//:result.json".split(' '), appConfig.path.toFile())
+      CommandRunner.run("bazel build @all_external_repositories//:result.json".split(' '), appConfig.workspaceRoot.toFile())
       workspaceFilePath.writeText(originalContent)
       deleteFile(tempFileForBzl)
-      val bazelPath = CommandRunner.run("bazel info output_base".split(' '), appConfig.path.toFile()).removeSuffix("\n")
+      val bazelPath = CommandRunner.run("bazel info output_base".split(' '), appConfig.workspaceRoot.toFile()).removeSuffix("\n")
       val resultFilePath = Path(bazelPath).resolve("external/all_external_repositories/result.json")
       if (!resultFilePath.exists()) {
         throw RuntimeException("Failed to find a file")
