+++ core/src/main/kotlin/org/virtuslab/bazelsteward/core/AppConfig.kt
@@ -1,10 +1,10 @@
 package org.virtuslab.bazelsteward.core
 
-import org.virtuslab.bazelsteward.core.common.GitClient.Companion.GitAuthor
+import org.virtuslab.bazelsteward.core.common.GitClient.GitAuthor
 import java.nio.file.Path
 
 data class AppConfig(
-  val path: Path,
+  val workspaceRoot: Path,
   val configPath: Path,
   val pushToRemote: Boolean,
   val baseBranch: String,
