+++ core/src/main/kotlin/org/virtuslab/bazelsteward/core/GitHostClient.kt
@@ -1,14 +1,10 @@
 package org.virtuslab.bazelsteward.core
 
-import org.virtuslab.bazelsteward.core.library.LibraryId
-import org.virtuslab.bazelsteward.core.library.Version
-
 data class PullRequest(val branch: GitBranch)
 
 interface GitHostClient {
   fun checkPrStatus(branch: GitBranch): PrStatus
   fun openNewPR(branch: GitBranch)
-  fun closePrs(library: LibraryId, filterNotVersion: Version? = null)
   fun getOpenPRs(): List<PullRequest>
   fun closePrs(pullRequest: List<PullRequest>)
 
@@ -20,7 +16,6 @@ interface GitHostClient {
     val stub = object : GitHostClient {
       override fun checkPrStatus(branch: GitBranch) = PrStatus.NONE
       override fun openNewPR(branch: GitBranch) {}
-      override fun closePrs(library: LibraryId, filterNotVersion: Version?) {}
       override fun getOpenPRs(): List<PullRequest> = emptyList()
       override fun closePrs(pullRequest: List<PullRequest>) {}
     }
