+++ core/src/main/kotlin/org/virtuslab/bazelsteward/core/common/CommandRunner.kt
@@ -7,6 +7,7 @@ import mu.KotlinLogging
 import java.io.File
 
 private val logger = KotlinLogging.logger {}
+
 class CommandRunner {
 
   companion object {
