+++ core/src/main/kotlin/org/virtuslab/bazelsteward/core/library/Library.kt
@@ -7,4 +7,6 @@ interface Library {
   val version: Version
   val versioningSchema: VersioningSchema
   val bumpingStrategy: BumpingStrategy
+  fun withVersioningSchema(schema: VersioningSchema): Library
+  fun withBumpingStrategy(strategy: BumpingStrategy): Library
 }
