+++ core/src/main/kotlin/org/virtuslab/bazelsteward/core/library/VersioningSchema.kt
@@ -15,7 +15,9 @@ sealed class VersioningSchema {
       }
     }
 
-    override fun equals(other: Any?): Boolean = (other is Regex) && other.regex.pattern == regex.pattern && other.regex.options == regex.options // for some reason Kotlin's Regex implementation compares instances by default
+    override fun equals(other: Any?): Boolean =
+      (other is Regex) && other.regex.pattern == regex.pattern && other.regex.options == regex.options // for some reason Kotlin's Regex implementation compares instances by default
+
     override fun hashCode(): Int = 42 * regex.pattern.hashCode() + regex.options.hashCode()
   }
 }
