+++ core/src/main/kotlin/org/virtuslab/bazelsteward/core/replacement/Heuristic.kt
@@ -1,10 +1,8 @@
 package org.virtuslab.bazelsteward.core.replacement
 
-import org.virtuslab.bazelsteward.core.common.BazelFileSearch
 import org.virtuslab.bazelsteward.core.common.FileChange
+import org.virtuslab.bazelsteward.core.common.TextFile
 import org.virtuslab.bazelsteward.core.common.UpdateLogic
-import org.virtuslab.bazelsteward.core.library.LibraryId
-import org.virtuslab.bazelsteward.core.library.Version
 
 data class LibraryUpdate(
   val updateSuggestion: UpdateLogic.UpdateSuggestion,
@@ -14,7 +12,7 @@ data class LibraryUpdate(
 interface Heuristic {
   val name: String
   fun apply(
-    files: List<BazelFileSearch.BazelFile>,
+    files: List<TextFile>,
     updateSuggestion: UpdateLogic.UpdateSuggestion
   ): LibraryUpdate?
 }
