+++ core/src/main/kotlin/org/virtuslab/bazelsteward/core/replacement/VersionOnlyHeuristic.kt
@@ -1,16 +1,14 @@
 package org.virtuslab.bazelsteward.core.replacement
 
-import org.virtuslab.bazelsteward.core.common.BazelFileSearch
 import org.virtuslab.bazelsteward.core.common.FileChange
+import org.virtuslab.bazelsteward.core.common.TextFile
 import org.virtuslab.bazelsteward.core.common.UpdateLogic
-import org.virtuslab.bazelsteward.core.library.LibraryId
-import org.virtuslab.bazelsteward.core.library.Version
 
 object VersionOnlyHeuristic : Heuristic {
   override val name: String = "version-only"
 
   override fun apply(
-    files: List<BazelFileSearch.BazelFile>,
+    files: List<TextFile>,
     updateSuggestion: UpdateLogic.UpdateSuggestion
   ): LibraryUpdate? {
     val currentVersion = updateSuggestion.currentLibrary.version.value
