+++ core/src/main/kotlin/org/virtuslab/bazelsteward/core/rules/RuleVersion.kt
@@ -3,6 +3,5 @@ package org.virtuslab.bazelsteward.core.rules
 import org.virtuslab.bazelsteward.core.library.Version
 
 data class RuleVersion(val url: String, val sha256: String, val tag: String) : Version() {
-  override val value: String
-    get() = tag
+  override val value: String = tag
 }
