+++ core/src/test/kotlin/org/virtuslab/bazelsteward/core/replacement/HeuristicTest.kt
@@ -4,9 +4,8 @@ import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
-import org.virtuslab.bazelsteward.core.common.BazelFileSearch
-import org.virtuslab.bazelsteward.core.common.FileChangeSuggestion
-import org.virtuslab.bazelsteward.core.common.FileUpdateSearch
+import org.virtuslab.bazelsteward.core.common.FileChange
+import org.virtuslab.bazelsteward.core.common.TextFile
 import org.virtuslab.bazelsteward.core.common.UpdateLogic
 import org.virtuslab.bazelsteward.core.config.BumpingStrategy
 import org.virtuslab.bazelsteward.core.library.SemanticVersion
@@ -22,11 +21,39 @@ class HeuristicTest {
   val correctPositionFor120: Int = 2263
   val correctPositionFor160: Int = 2464
 
-  data class TestBazelFile(override val path: Path, override val content: String) : BazelFileSearch.BazelFile
+  data class TestTextFile(override val path: Path, override val content: String) : TextFile
 
-  private fun getBazelFile(fileName: String): BazelFileSearch.BazelFile {
+  private fun loadTextFileFromResources(fileName: String): TextFile {
     val url = this::class.java.classLoader.getResource(fileName)!!
-    return TestBazelFile(Path(url.path), url.readText())
+    return TestTextFile(Path(url.path), url.readText())
+  }
+
+  private val files = listOf(loadTextFileFromResources("WORKSPACE.bzlignore"))
+
+  private val suggester = FileChangeSuggester()
+
+  private fun library(group: String, artifact: String, version: String) =
+    MavenCoordinates(
+      MavenLibraryId(group, artifact),
+      version(version),
+      VersioningSchema.SemVer,
+      BumpingStrategy.Default
+    )
+
+  private fun version(version: String) = SemanticVersion.fromString(version)!!
+
+  private val allHeuristics = listOf(WholeLibraryHeuristic, VersionOnlyHeuristic).toTypedArray()
+
+  private fun run(
+    library: MavenCoordinates,
+    version: SemanticVersion,
+    vararg heuristics: Heuristic = allHeuristics
+  ): FileChange? {
+    return suggester.suggestChanges(
+      files,
+      UpdateLogic.UpdateSuggestion(library, version),
+      heuristics.toList()
+    )?.fileChanges?.firstOrNull()
   }
 
   @Nested
@@ -35,97 +62,52 @@ class HeuristicTest {
 
     @Test
     fun `should return correct position offset`() {
-      val fileUpdateSearch = FileUpdateSearch()
-      val lib = MavenCoordinates(
-        MavenLibraryId("com.7theta", "utilis"),
-        SemanticVersion(2, 3, 5, "", ""),
-        VersioningSchema.SemVer,
-        BumpingStrategy.Default
-      )
-      val suggestedVersion = SemanticVersion(2, 4, 0, "", "")
-      val updateSuggestion = UpdateLogic.UpdateSuggestion(lib, suggestedVersion)
+      val library = library("com.7theta", "utilis", "2.3.5")
+      val suggestedVersion = version("2.4.0")
 
-      val files = listOf(getBazelFile("WORKSPACE.bzlignore"))
+      val result = run(library, suggestedVersion)
 
-      val result: List<FileChangeSuggestion> =
-        fileUpdateSearch.searchBuildFiles(files, listOf(updateSuggestion))
-      result[0].position shouldBe correctPositionFor235
+      result?.offset shouldBe correctPositionFor235
     }
 
     @Test
     fun `should return correct position offset with wrong artifact`() {
-      val fileUpdateSearch = FileUpdateSearch()
-      val lib = MavenCoordinates(
-        MavenLibraryId("com.10theta", "utilis"),
-        SemanticVersion(2, 3, 5, "", ""),
-        VersioningSchema.SemVer,
-        BumpingStrategy.Default
-      )
-      val suggestedVersion = SemanticVersion(2, 4, 0, "", "")
-      val updateSuggestion = UpdateLogic.UpdateSuggestion(lib, suggestedVersion)
+      val library = library("com.10theta", "utilis", "2.3.5")
+      val suggestedVersion = version("2.4.0")
 
-      val files = listOf(getBazelFile("WORKSPACE.bzlignore"))
+      val result = run(library, suggestedVersion)
 
-      val result: List<FileChangeSuggestion> =
-        fileUpdateSearch.searchBuildFiles(files, listOf(updateSuggestion))
-      result[0].position shouldBe correctPositionFor235
+      result?.offset shouldBe correctPositionFor235
     }
 
     @Test
     fun `should return correct position offset without artifact`() {
-      val fileUpdateSearch = FileUpdateSearch()
-      val lib = MavenCoordinates(
-        MavenLibraryId("", ""),
-        SemanticVersion(2, 3, 5, "", ""),
-        VersioningSchema.SemVer,
-        BumpingStrategy.Default
-      )
-      val suggestedVersion = SemanticVersion(2, 4, 0, "", "")
-      val updateSuggestion = UpdateLogic.UpdateSuggestion(lib, suggestedVersion)
+      val library = library("", "", "2.3.5")
+      val suggestedVersion = version("2.4.0")
 
-      val files = listOf(getBazelFile("WORKSPACE.bzlignore"))
+      val result = run(library, suggestedVersion)
 
-      val result: List<FileChangeSuggestion> =
-        fileUpdateSearch.searchBuildFiles(files, listOf(updateSuggestion))
-      result[0].position shouldBe correctPositionFor235
+      result?.offset shouldBe correctPositionFor235
     }
 
     @Test
     fun `should return correct position offset for artifact with version in variable`() {
-      val fileUpdateSearch = FileUpdateSearch()
-      val lib = MavenCoordinates(
-        MavenLibraryId("io.grpc", "grpc-core"),
-        SemanticVersion(1, 2, 0, "", ""),
-        VersioningSchema.SemVer,
-        BumpingStrategy.Default
-      )
-      val suggestedVersion = SemanticVersion(2, 4, 0, "", "")
-      val updateSuggestion = UpdateLogic.UpdateSuggestion(lib, suggestedVersion)
+      val library = library("io.grpc", "grpc-core", "1.2.0")
+      val suggestedVersion = version("2.4.0")
 
-      val files = listOf(getBazelFile("WORKSPACE.bzlignore"))
+      val result = run(library, suggestedVersion)
 
-      val result: List<FileChangeSuggestion> =
-        fileUpdateSearch.searchBuildFiles(files, listOf(updateSuggestion))
-      result[0].position shouldBe correctPositionFor120
+      result?.offset shouldBe correctPositionFor120
     }
 
     @Test
     fun `should return correct position offset when two libraries have same version`() {
-      val fileUpdateSearch = FileUpdateSearch()
-      val lib = MavenCoordinates(
-        MavenLibraryId("org.jetbrains.kotlinx", "kotlinx-coroutines-core"),
-        SemanticVersion(1, 6, 0, "", ""),
-        VersioningSchema.SemVer,
-        BumpingStrategy.Default
-      )
-      val suggestedVersion = SemanticVersion(2, 4, 0, "", "")
-      val updateSuggestion = UpdateLogic.UpdateSuggestion(lib, suggestedVersion)
+      val library = library("org.jetbrains.kotlinx", "kotlinx-coroutines-core", "1.6.0")
+      val suggestedVersion = version("2.4.0")
 
-      val files = listOf(getBazelFile("WORKSPACE.bzlignore"))
+      val result = run(library, suggestedVersion)
 
-      val result: List<FileChangeSuggestion> =
-        fileUpdateSearch.searchBuildFiles(files, listOf(updateSuggestion))
-      result[0].position shouldBe correctPositionFor160
+      result?.offset shouldBe correctPositionFor160
     }
   }
 
@@ -135,92 +117,52 @@ class HeuristicTest {
 
     @Test
     fun `should return correct position`() {
-      val wholeLibraryHeuristic = WholeLibraryHeuristic
-      val lib = MavenCoordinates(
-        MavenLibraryId("com.7theta", "utilis"),
-        SemanticVersion(2, 3, 5, "", ""),
-        VersioningSchema.SemVer,
-        BumpingStrategy.Default
-      )
-      val suggestedVersion = SemanticVersion(2, 4, 0, "", "")
-      val updateSuggestion = UpdateLogic.UpdateSuggestion(lib, suggestedVersion)
+      val library = library("com.7theta", "utilis", "2.3.5")
+      val suggestedVersion = version("2.4.0")
 
-      val files = listOf(getBazelFile("WORKSPACE.bzlignore"))
+      val result = run(library, suggestedVersion, WholeLibraryHeuristic)
 
-      val result = wholeLibraryHeuristic.apply(files, updateSuggestion)!!
-      result.position shouldBe correctPositionFor235
+      result?.offset shouldBe correctPositionFor235
     }
 
     @Test
     fun `should return null with wrong artifact`() {
-      val wholeLibraryHeuristic = WholeLibraryHeuristic
-      val lib = MavenCoordinates(
-        MavenLibraryId("com.10theta", "utilis"),
-        SemanticVersion(2, 3, 5, "", ""),
-        VersioningSchema.SemVer,
-        BumpingStrategy.Default
-      )
-      val suggestedVersion = SemanticVersion(2, 4, 0, "", "")
-      val updateSuggestion = UpdateLogic.UpdateSuggestion(lib, suggestedVersion)
+      val library = library("com.10theta", "utilis", "2.3.5")
+      val suggestedVersion = version("2.4.0")
 
-      val files = listOf(getBazelFile("WORKSPACE.bzlignore"))
+      val result = run(library, suggestedVersion, WholeLibraryHeuristic)
 
-      val result = wholeLibraryHeuristic.apply(files, updateSuggestion)
       result shouldBe null
     }
 
     @Test
     fun `should return correct position without artifact`() {
-      val wholeLibraryHeuristic = WholeLibraryHeuristic
-      val lib = MavenCoordinates(
-        MavenLibraryId("", ""),
-        SemanticVersion(2, 3, 5, "", ""),
-        VersioningSchema.SemVer,
-        BumpingStrategy.Default
-      )
-      val suggestedVersion = SemanticVersion(2, 4, 0, "", "")
-      val updateSuggestion = UpdateLogic.UpdateSuggestion(lib, suggestedVersion)
+      val library = library("", "", "2.3.5")
+      val suggestedVersion = version("2.4.0")
 
-      val files = listOf(getBazelFile("WORKSPACE.bzlignore"))
+      val result = run(library, suggestedVersion, WholeLibraryHeuristic)
 
-      val result = wholeLibraryHeuristic.apply(files, updateSuggestion)!!
-      result.position shouldBe correctPositionFor235
+      result?.offset shouldBe correctPositionFor235
     }
 
     @Test
     fun `should return null for artifact with version in variable`() {
-      val wholeLibraryHeuristic = WholeLibraryHeuristic
-      val lib = MavenCoordinates(
-        MavenLibraryId("io.grpc", "grpc-core"),
-        SemanticVersion(1, 2, 0, "", ""),
-        VersioningSchema.SemVer,
-        BumpingStrategy.Default
-      )
-      val suggestedVersion = SemanticVersion(2, 4, 0, "", "")
-      val updateSuggestion = UpdateLogic.UpdateSuggestion(lib, suggestedVersion)
+      val library = library("io.grpc", "grpc-core", "1.2.0")
+      val suggestedVersion = version("2.4.0")
 
-      val files = listOf(getBazelFile("WORKSPACE.bzlignore"))
+      val result = run(library, suggestedVersion, WholeLibraryHeuristic)
 
-      val result = wholeLibraryHeuristic.apply(files, updateSuggestion)
       result shouldBe null
     }
 
     @Test
     fun `should return correct position offset when two libraries have same version`() {
-      val wholeLibraryHeuristic = WholeLibraryHeuristic
-      val lib = MavenCoordinates(
-        MavenLibraryId("org.jetbrains.kotlinx", "kotlinx-coroutines-core"),
-        SemanticVersion(1, 6, 0, "", ""),
-        VersioningSchema.SemVer,
-        BumpingStrategy.Default
-      )
-      val suggestedVersion = SemanticVersion(2, 4, 0, "", "")
-      val updateSuggestion = UpdateLogic.UpdateSuggestion(lib, suggestedVersion)
+      val library = library("org.jetbrains.kotlinx", "kotlinx-coroutines-core", "1.6.0")
+      val suggestedVersion = version("2.4.0")
 
-      val files = listOf(getBazelFile("WORKSPACE.bzlignore"))
+      val result = run(library, suggestedVersion, WholeLibraryHeuristic)
 
-      val result = wholeLibraryHeuristic.apply(files, updateSuggestion)!!
-      result.position shouldBe correctPositionFor160
+      result?.offset shouldBe correctPositionFor160
     }
   }
 
@@ -230,91 +172,51 @@ class HeuristicTest {
 
     @Test
     fun `should return correct position offset `() {
-      val versionOnlyHeuristic = VersionOnlyHeuristic
-      val lib = MavenCoordinates(
-        MavenLibraryId("com.7theta", "utilis"),
-        SemanticVersion(2, 3, 5, "", ""),
-        VersioningSchema.SemVer,
-        BumpingStrategy.Default
-      )
-      val suggestedVersion = SemanticVersion(2, 4, 0, "", "")
-      val updateSuggestion = UpdateLogic.UpdateSuggestion(lib, suggestedVersion)
+      val library = library("com.7theta", "utilis", "2.3.5")
+      val suggestedVersion = version("2.4.0")
 
-      val files = listOf(getBazelFile("WORKSPACE.bzlignore"))
+      val result = run(library, suggestedVersion, VersionOnlyHeuristic)
 
-      val result = versionOnlyHeuristic.apply(files, updateSuggestion)!!
-      result.fileChanges[0].offset shouldBe correctPositionFor235
+      result?.offset shouldBe correctPositionFor235
     }
 
     @Test
     fun `should return correct position offset with wrong artifact`() {
-      val versionOnlyHeuristic = VersionOnlyHeuristic
-      val lib = MavenCoordinates(
-        MavenLibraryId("com.10theta", "utilis"),
-        SemanticVersion(2, 3, 5, "", ""),
-        VersioningSchema.SemVer,
-        BumpingStrategy.Default
-      )
-      val suggestedVersion = SemanticVersion(2, 4, 0, "", "")
-      val updateSuggestion = UpdateLogic.UpdateSuggestion(lib, suggestedVersion)
+      val library = library("com.10theta", "utilis", "2.3.5")
+      val suggestedVersion = version("2.4.0")
 
-      val files = listOf(getBazelFile("WORKSPACE.bzlignore"))
+      val result = run(library, suggestedVersion, VersionOnlyHeuristic)
 
-      val result = versionOnlyHeuristic.apply(files, updateSuggestion)!!
-      result.position shouldBe correctPositionFor235
+      result?.offset shouldBe correctPositionFor235
     }
 
     @Test
     fun `should return correct position offset without artifact`() {
-      val versionOnlyHeuristic = VersionOnlyHeuristic()
-      val lib = MavenCoordinates(
-        MavenLibraryId("", ""),
-        SemanticVersion(2, 3, 5, "", ""),
-        VersioningSchema.SemVer,
-        BumpingStrategy.Default
-      )
-      val suggestedVersion = SemanticVersion(2, 4, 0, "", "")
-      val updateSuggestion = UpdateLogic.UpdateSuggestion(lib, suggestedVersion)
+      val library = library("", "", "2.3.5")
+      val suggestedVersion = version("2.4.0")
 
-      val files = listOf(getBazelFile("WORKSPACE.bzlignore"))
+      val result = run(library, suggestedVersion, VersionOnlyHeuristic)
 
-      val result = versionOnlyHeuristic.apply(files, updateSuggestion)!!
-      result.position shouldBe correctPositionFor235
+      result?.offset shouldBe correctPositionFor235
     }
 
     @Test
     fun `should return correct position for artifact with version in variable`() {
-      val versionOnlyHeuristic = VersionOnlyHeuristic()
-      val lib = MavenCoordinates(
-        MavenLibraryId("io.grpc", "grpc-core"),
-        SemanticVersion(1, 2, 0, "", ""),
-        VersioningSchema.SemVer,
-        BumpingStrategy.Default
-      )
-      val suggestedVersion = SemanticVersion(2, 4, 0, "", "")
-      val updateSuggestion = UpdateLogic.UpdateSuggestion(lib, suggestedVersion)
+      val library = library("io.grpc", "grpc-core", "1.2.0")
+      val suggestedVersion = version("2.4.0")
 
-      val files = listOf(getBazelFile("WORKSPACE.bzlignore"))
+      val result = run(library, suggestedVersion, VersionOnlyHeuristic)
 
-      val result = versionOnlyHeuristic.apply(files, updateSuggestion)!!
-      result.position shouldBe correctPositionFor120
+      result?.offset shouldBe correctPositionFor120
     }
 
     @Test
     fun `should return null when two libraries have same version`() {
-      val versionOnlyHeuristic = VersionOnlyHeuristic()
-      val lib = MavenCoordinates(
-        MavenLibraryId("org.jetbrains.kotlinx", "kotlinx-coroutines-core"),
-        SemanticVersion(1, 6, 0, "", ""),
-        VersioningSchema.SemVer,
-        BumpingStrategy.Default
-      )
-      val suggestedVersion = SemanticVersion(2, 4, 0, "", "")
-      val updateSuggestion = UpdateLogic.UpdateSuggestion(lib, suggestedVersion)
+      val library = library("org.jetbrains.kotlinx", "kotlinx-coroutines-core", "1.6.0")
+      val suggestedVersion = version("2.4.0")
 
-      val files = listOf(getBazelFile("WORKSPACE.bzlignore"))
+      val result = run(library, suggestedVersion, VersionOnlyHeuristic)
 
-      val result = versionOnlyHeuristic.apply(files, updateSuggestion)
       result shouldBe null
     }
   }
@@ -325,45 +227,25 @@ class HeuristicTest {
 
     @Test
     fun `should return same position offset for WholeVersionHeuristic and VersionHeuristic`() {
-      val lib = MavenCoordinates(
-        MavenLibraryId("com.7theta", "utilis"),
-        SemanticVersion(2, 3, 5, "", ""),
-        VersioningSchema.SemVer,
-        BumpingStrategy.Default
-      )
-      val suggestedVersion = SemanticVersion(2, 4, 0, "", "")
-      val updateSuggestion = UpdateLogic.UpdateSuggestion(lib, suggestedVersion)
+      val library = library("com.7theta", "utilis", "2.3.5")
+      val suggestedVersion = version("2.4.0")
 
-      val files = listOf(getBazelFile("WORKSPACE.bzlignore"))
+      val result1 = run(library, suggestedVersion, WholeLibraryHeuristic)
+      val result2 = run(library, suggestedVersion, VersionOnlyHeuristic)
 
-      val wholeLibraryHeuristic = WholeLibraryHeuristic()
-      val result1 = wholeLibraryHeuristic.apply(files, updateSuggestion)!!
-      val versionOnlyHeuristic = VersionOnlyHeuristic()
-      val result2 = versionOnlyHeuristic.apply(files, updateSuggestion)!!
-
-      result1.position shouldBe result2.position
+      result1?.offset shouldBe result2?.offset
     }
 
     @Test
     fun `should return different position offset for WholeVersionHeuristic and VersionHeuristic`() {
-      val lib = MavenCoordinates(
-        MavenLibraryId("com.10theta", "utilis"),
-        SemanticVersion(2, 3, 5, "", ""),
-        VersioningSchema.SemVer,
-        BumpingStrategy.Default
-      )
-      val suggestedVersion = SemanticVersion(2, 4, 0, "", "")
-      val updateSuggestion = UpdateLogic.UpdateSuggestion(lib, suggestedVersion)
+      val library = library("com.10theta", "utilis", "2.3.5")
+      val suggestedVersion = version("2.4.0")
 
-      val files = listOf(getBazelFile("WORKSPACE.bzlignore"))
-
-      val wholeLibraryHeuristic = WholeLibraryHeuristic()
-      val result1 = wholeLibraryHeuristic.apply(files, updateSuggestion)
-      val versionOnlyHeuristic = VersionOnlyHeuristic()
-      val result2 = versionOnlyHeuristic.apply(files, updateSuggestion)!!
+      val result1 = run(library, suggestedVersion, WholeLibraryHeuristic)
+      val result2 = run(library, suggestedVersion, VersionOnlyHeuristic)
 
       result1 shouldBe null
-      result2.position shouldBe correctPositionFor235
+      result2?.offset shouldBe correctPositionFor235
     }
   }
 }
